requirements.txt
BTrees==4.0.8
DateTime==4.0.1
dateutils==0.6.6
Django==1.6
django-model-utils==2.0.3
django-appconf==1.0.1
django-compressor==1.6
django-dajax==0.9.2
django-dajaxice==0.6
django-debug-toolbar==1.2.2
django-extensions==1.3.3
Django-Select2==4.2.2
django-sorl-hacks==0.1.10
django-tinymce==1.5.2
Django-tinymce-filebrowser==0.2.1
djangocms-stacks==0.2.3
docutils==0.12
easy-thumbnails==1.4
feedgen==0.3.2
fpconst==0.7.2
gunicorn==19.4.5
lxml==3.5.0
MySQL-python==1.2.5
mysqldbda==1.0.2
persistent==4.0.8
Pillow==2.6.1
python-dateutil==2.4.2
pytz==2014.10
rcssmin==1.0.6
rjsmin==1.0.12
simple-json==1.1
simplejson==3.6.5
six==1.8.0
SOAPpy==0.12.1
sorl-thumbnail==11.12
South==1.0.1
sqlparse==0.1.14
transaction==1.4.3
virtualenv==1.11.6
wstools==0.4.3
zc.lockfile==1.1.0
ZConfig==3.0.4
zdaemon==4.0.0
ZEO==4.0.0
ZODB==4.0.1
ZODB3==3.11.0
zope.app.form==4.0.2
zope.browser==2.0.2
zope.browsermenu==4.1.0a1
zope.browserpage==4.1.0a1
zope.component==4.2.1
zope.configuration==4.0.3
zope.container==4.0.0
zope.contenttype==4.0.1
zope.datetime==4.0.0
zope.dottedname==4.0.1
zope.event==4.0.3
zope.exceptions==4.0.7
zope.filerepresentation==4.0.2
zope.formlib==4.3.0a2
zope.i18n==4.0.0a4
zope.i18nmessageid==4.0.3
zope.interface==4.1.2
zope.lifecycleevent==4.0.3
zope.location==4.0.3
zope.pagetemplate==4.0.4
zope.proxy==4.1.4
zope.publisher==4.0.0a4
zope.rdb==3.5.0
zope.schema==4.4.2
zope.security==4.0.1
zope.size==4.0.1
zope.tal==4.0.0
zope.tales==4.0.2
zope.thread==3.4
zope.traversing==4.0.0
Wand==0.4.4